+++ src/main/java/com/homestore/forum/ForumRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ForumRepository extends JpaRepository<Forum, Integer> {
+public interface ForumRepository extends JpaRepository<Forum, Long> {
     
     Optional<Forum> findByName(ForumNameEnum name);
 }
